requirements.txt
opencv-python
PySide2
numba
ninja
matplotlib
scipy
lap
sklearn
torchsummary
pycuda
pyzmq